+++ app/assets/composables/useRolePermissionsApi.ts
@@ -1,16 +1,14 @@
 import { ref } from 'vue'
 import axios from 'axios'
 import type { ApiErrorResponse } from '@userfrosting/sprinkle-core/interfaces'
-import type { PermissionInterface } from '@userfrosting/sprinkle-account/interfaces'
-import type { PermissionSprunjeResponse, RolePermissionsSprunjeResponse } from '../interfaces'
+import type { RolePermissionsSprunjeResponse } from '../interfaces'
 
 /**
  * API used to fetch a match between all available permissions and the role's
  * permissions, in a single component
  *
- * This API is tied to the `PermissionsSprunje` and `RolePermissionsSprunje` API,
- *  accessed at the GET `/api/permissions` and
- * `/api/roles/r/{slug}/permissions` endpoints.
+ * This API is tied to the `RolePermissionsSprunje` API, accessed at
+ * the GET `/api/roles/r/{slug}/permissions` endpoint.
  *
  * This composable accept a {roleSlug} to select the permissions of a specific
  * role.
@@ -19,41 +17,13 @@ export function useRolePermissionsApi() {
     const loading = ref<boolean>(false)
     const error = ref<ApiErrorResponse | null>()
     const selected = ref<number[]>([])
-    const permissions = ref<PermissionInterface[]>([])
 
-    // Step 1 - Fetch all permissions
+    // Fetch role's permissions and match them with the permissions
     async function fetch(roleSlug: string) {
-        loading.value = true
-        axios
-            .get<PermissionSprunjeResponse>('/api/permissions')
-            .then((response) => {
-                permissions.value = response.data.rows
-                fetchRolePermissions(roleSlug)
-            })
-            .catch((err) => {
-                loading.value = false
-                error.value = err.response.data
-            })
-    }
-
-    // Step 2 - Fetch role permissions and match them with the permissions
-    async function fetchRolePermissions(roleSlug: string) {
         axios
             .get<RolePermissionsSprunjeResponse>('/api/roles/r/' + roleSlug + '/permissions')
             .then((response) => {
-                // Empty the selected array
-                selected.value.splice(0)
-
-                // Match the permissions with the role permissions
-                const rolePermissions: PermissionInterface[] = response.data.rows
-                rolePermissions.forEach((rolePermission) => {
-                    const record = permissions.value.find(
-                        (element) => element.id === rolePermission.id
-                    )
-                    if (record) {
-                        selected.value.push(rolePermission.id)
-                    }
-                })
+                selected.value = response.data.rows.map((permission) => permission.id)
             })
             .catch((err) => {
                 error.value = err.response.data
@@ -63,5 +33,5 @@ export function useRolePermissionsApi() {
             })
     }
 
-    return { error, loading, fetch, selected, permissions }
+    return { error, loading, fetch, selected }
 }