+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Simplified `useRolePermissionsApi` and `useUserRolesApi` to be used with the Sprunje API, instead of a separate API.
 
 ## [6.0.0](https://github.com/userfrosting/sprinkle-admin/compare/6.0.0-rc.5...6.0.0) - 2026-06-12
 - No changes.